requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
trafilatura==1.12.2
readability-lxml==0.8.1
gTTS==2.5.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
